libdmi: update intel_0x94 table description for Broxton
Intel 0x94 tables now define a new key-value/key-type-value
format, which is being first suported on Broxton, in libdmi.
Support for the new tables is currently offered on newer
platforms and is applied based on the platforms name which is
checked in the SYSTEM_INFORMATION table.

A test scenario for the key/value table implementation is added.

Change-Id: I8df88c9b1bf5197f3c318cd497c8dc2a6ebb55e2
Tracked-On: OAM-494
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 403137